Added basic character selection and button interfacing
Went through and put in manual button navigation so it may meet our odd
2 player req.

Added a basic character select off of the play button.   this is set up
so we can make copies of the button for various other characters (but
for now only features the link animation.)

There is now a cross scene object that can carry information back and
forth from scene to scene.

Added my test scene to the build properties.