#3 Added 2 more prefabs that look like circuit boards
I would still like to see if a few people would be interested in doing
some tiles. Variety would be helpful.  If you want to throw something
together in MagicaVoxel the dimensions I worked in was (100 100 x.)
Build as high as you want. Maybe even make one that is all in one.