Fixed the data points for stats.
#31 First time I sit down to work on this thing in two weeks and I
figure it out in 10 minutes. . .

In short, when the stats were being calculated they only calculated half
the points( mostly the ones that are 0's because it's keeping track when
the countdown is going) because it was referenceing the size of the
array list (which was shrinking every single call). I could be wrong, in
which case we can change it back, but another issue was it was
destroying the list because it was being passed by reference.

Anyway, if I did something stupid. Just revert this change. But data is
being displayed correctly now.